+++ src/__tests__/scenes/FioAddressSettingsScene.test.tsx
@@ -4,7 +4,7 @@ import { createRenderer } from 'react-test-renderer/shallow'
 
 import { FioAddressSettingsComponent } from '../../components/scenes/Fio/FioAddressSettingsScene'
 import { getTheme } from '../../components/services/ThemeContext'
-import { fakeSceneProps } from '../../util/fake/fakeSceneProps'
+import { fakeEdgeAppSceneProps } from '../../util/fake/fakeSceneProps'
 
 describe('FioAddressSettingsComponent', () => {
   it('should render with loading props', () => {
@@ -24,7 +24,7 @@ describe('FioAddressSettingsComponent', () => {
 
     const actual = renderer.render(
       <FioAddressSettingsComponent
-        {...fakeSceneProps('fioAddressSettings', {
+        {...fakeEdgeAppSceneProps('fioAddressSettings', {
           walletId: fakeWallet.id,
           fioAddressName: 'MyFioAddress'
         })}