+++ src/components/scenes/MigrateWalletCompletionScene.tsx
@@ -13,7 +13,7 @@ import { useHandler } from '../../hooks/useHandler'
 import { useWatch } from '../../hooks/useWatch'
 import { lstrings } from '../../locales/strings'
 import { useSelector } from '../../types/reactRedux'
-import { EdgeSceneProps } from '../../types/routerTypes'
+import { EdgeAppSceneProps } from '../../types/routerTypes'
 import { getWalletName } from '../../util/CurrencyWalletHelpers'
 import { SceneWrapper } from '../common/SceneWrapper'
 import { showError } from '../services/AirshipInstance'
@@ -27,7 +27,7 @@ export interface MigrateWalletCompletionParams {
   migrateWalletList: MigrateWalletItem[]
 }
 
-interface Props extends EdgeSceneProps<'migrateWalletCompletion'> {}
+interface Props extends EdgeAppSceneProps<'migrateWalletCompletion'> {}
 
 interface MigrateWalletTokenItem extends MigrateWalletItem {
   tokenId: string
@@ -239,7 +239,7 @@ const MigrateWalletCompletionComponent = (props: Props) => {
           label={!done ? undefined : lstrings.string_done_cap}
           type="secondary"
           marginRem={[0, 0, 1]}
-          onPress={() => navigation.navigate('walletsTab', { screen: 'walletList' })}
+          onPress={() => navigation.navigate('edgeTabs', { screen: 'walletsTab', params: { screen: 'walletList' } })}
         />
       </View>
     )