+++ src/components/scenes/MigrateWalletSelectCryptoScene.tsx
@@ -9,7 +9,7 @@ import { useWatch } from '../../hooks/useWatch'
 import { lstrings } from '../../locales/strings'
 import { WalletCreateItem } from '../../selectors/getCreateWalletList'
 import { useSelector } from '../../types/reactRedux'
-import { EdgeSceneProps } from '../../types/routerTypes'
+import { EdgeAppSceneProps } from '../../types/routerTypes'
 import { getCurrencyCode, isKeysOnlyPlugin } from '../../util/CurrencyInfoHelpers'
 import { getWalletName } from '../../util/CurrencyWalletHelpers'
 import { zeroString } from '../../util/utils'
@@ -25,7 +25,7 @@ export interface MigrateWalletSelectCryptoParams {
   preSelectedWalletIds?: string[]
 }
 
-interface Props extends EdgeSceneProps<'migrateWalletSelectCrypto'> {}
+interface Props extends EdgeAppSceneProps<'migrateWalletSelectCrypto'> {}
 
 export interface MigrateWalletItem extends WalletCreateItem {
   createWalletIds: [string]